# docs/api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Music Gen AI API
  version: 1.0.0
  description: |
    Production-ready AI music generation API that creates high-quality music from text prompts using transformer architecture with EnCodec audio tokenization.

    ## Features
    - **Text-to-Music Generation**: Generate music from natural language descriptions
    - **Multiple Model Support**: Various model sizes for different use cases
    - **Real-time Streaming**: Stream music generation progress
    - **High-Quality Audio**: 32kHz sampling rate with EnCodec compression
    - **Rate Limiting**: Tier-based usage controls
    - **Authentication**: JWT-based secure access
    - **Monitoring**: Comprehensive metrics and health endpoints

    ## Getting Started
    1. Sign up for an API key at [https://musicgen.ai/signup](https://musicgen.ai/signup)
    2. Authenticate using JWT tokens or API keys
    3. Start generating music with simple text prompts
    4. Monitor your usage and billing in the dashboard

    ## Support
    - **Documentation**: [https://docs.musicgen.ai](https://docs.musicgen.ai)
    - **Support**: [support@musicgen.ai](mailto:support@musicgen.ai)
    - **Status Page**: [https://status.musicgen.ai](https://status.musicgen.ai)

  termsOfService: https://musicgen.ai/terms
  contact:
    name: Music Gen AI Support
    url: https://musicgen.ai/support
    email: support@musicgen.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - ApiKeyAuth: []
  - JwtAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: |
        Authenticate user credentials and receive JWT access token.
        
        **Rate Limit**: 5 requests per minute per IP
        
        **Security**: Passwords are hashed with bcrypt, failed attempts are logged
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              standard_login:
                summary: Standard email/password login
                value:
                  email: user@example.com
                  password: secure_password123
              api_key_login:
                summary: API key authentication
                value:
                  api_key: sk_test_1234567890abcdef
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
                expires_in: 3600
                refresh_token: refresh_1234567890abcdef
                user:
                  id: user_1234567890
                  email: user@example.com
                  tier: pro
                  usage:
                    generations_remaining: 450
                    reset_date: "2024-01-16T00:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            example:
              refresh_token: refresh_1234567890abcdef
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Include your API key in the `X-API-Key` header.
        
        Example: `X-API-Key: sk_live_1234567890abcdef`
        
        Get your API key from the [dashboard](https://musicgen.ai/dashboard).

    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. Include the token in the `Authorization` header.
        
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
        
        Obtain tokens via the `/auth/login` endpoint.
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: User password
          example: secure_password123
        api_key:
          type: string
          pattern: '^sk_[a-z]+_[a-zA-Z0-9]{32}$'
          description: API key (alternative to email/password)
          example: sk_test_1234567890abcdef1234567890abcdef
      oneOf:
        - required: [email, password]
        - required: [api_key]
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: Valid refresh token
          example: refresh_1234567890abcdef
      required: [refresh_token]
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [bearer]
          description: Token type
        expires_in:
          type: integer
          description: Token expiration time in seconds
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
        user:
          $ref: '#/components/schemas/User'
      required: [access_token, token_type, expires_in, user]
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: user_1234567890
        email:
          type: string
          format: email
          description: User email address
        tier:
          type: string
          enum: [free, pro, enterprise]
          description: User subscription tier
        usage:
          type: object
          properties:
            generations_remaining:
              type: integer
              description: Remaining generations in current period
            reset_date:
              type: string
              format: date-time
              description: When usage resets
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: "The prompt field is required"
        details:
          type: string
          description: Additional error details
        request_id:
          type: string
          description: Unique request identifier for debugging
          example: req_1234567890abcdef
      required: [code, message]
  responses:
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_auth:
              summary: Missing authentication
              value:
                code: AUTHENTICATION_REQUIRED
                message: "Authentication is required to access this resource"
                request_id: req_1234567890abcdef
            invalid_token:
              summary: Invalid token
              value:
                code: INVALID_TOKEN
                message: "The provided authentication token is invalid or expired"
                request_id: req_1234567890abcdef
            expired_token:
              summary: Expired token
              value:
                code: TOKEN_EXPIRED
                message: "Authentication token has expired. Please refresh your token."
                request_id: req_1234567890abcdef
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the rate limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            rate_limit_exceeded:
              summary: Rate limit exceeded
              value:
                code: RATE_LIMIT_EXCEEDED
                message: "Rate limit exceeded. Please wait before making another request."
                details: "You have exceeded the rate limit of 10 requests per minute"
                request_id: req_1234567890abcdef
...